Fix clang -Wconversion warnings in .hh files
The previous patch made compilation with gcc -Wconversion warning free.
Now follows a series of patches to do the same for clang.

'-Wconversion' actually enables a group of warnings. Though the details
differ between gcc and clang. In particular in gcc (in c++ code)
-Wconversion does NOT enable -Wsign-conversion, but in clang it does. At
this point the code is not -Wsign-conversion clean. So for now we'll
also disable this (sub)warning in clang (with -Wno-sign-conversion).

Even with -Wno-sign-conversion the code is not warning-free in clang.
By far most of the remaining warnings fall in the category: "a smarter
compiler could have seen that the expression can't overflow". Note that
this does not necessarily mean that gcc is smarter than clang (in this
aspect). There could also be false-positive warnings generated by gcc but
not by clang.

This patch fixes the clang warnings in .hh files. Those are repeated a
lot. So this already fixes the majority of the warning output.

+++ src/input/UnicodeKeymap.hh
@@ -86,7 +86,7 @@ public:
 	  */
 	[[nodiscard]] constexpr uint8_t getMask() const {
 		assert(isValid());
-		return 1 << getColumn();
+		return uint8_t(1 << getColumn());
 	}
 
 	[[nodiscard]] constexpr bool operator==(const KeyMatrixPosition&) const = default;
